New Feature: Benchmark Core-dump generation
Added a new feature to generate core-dump of when application/benchmark running
under VM crashes. It uses ptlcall to transfer core-dump generated in VM to host
and save it under 'appname-core' file. To enable this feature take a look at
new file 'core-dump-handler.c' in 'ptlsim/tools'.

This feature is added to aid debugging of applcations that crashes in VM while
running in simulation mode.

Signed-off-by: Avadh Patel <avadh4all@gmail.com>