New Feature: Synchronization between simulations
Adding a new feature 'sync' that will synchronize between multiple concurrent
simulations.  This feature is useful when you are running two or more
simulation instatnces that communicate with eachother via network. For example
running web-server and client in different simulation instances and you want to
run them with similar simulation speed.

Options are:
  -sync N          # Enable syncronization among N simulation instances
  -sync-master     # declare this instance as Master (only one can be master)
  -sync-interval N # Number of cycles to execute before each sync

Signed-off-by: Avadh Patel <avadh4all@gmail.com>